Update main.go
A simple hangman game I made using Golang. I used this as an exercise to test my knowledge of the various basic components of the Go language from loops, condition statements, package handling, functions and the range keyword. It was a very fun to work on and I look forward to throwing myself at more Golang projects as I continue to improve my knowledge and programming ability in this language.